requirements.txt
langchain
langchain-openai
python-dotenv
requests
fastapi
uvicorn[standard]